% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVAShiny.R
\name{ANOVAShiny}
\alias{ANOVAShiny}
\title{Launch 'ANOVAShiny' Interface}
\usage{
ANOVAShiny()
}
\value{
Nothing
}
\description{
ANOVAShiny() loads interactive user interface built using R 'shiny'.
}
\details{
Start ANOVAShiny

The interactive user interface is to provide an easy way for anova analysis and downloading relevant plots.
}
\examples{
if(interactive()){
library(rmarkdown)
ANOVAShiny()
}
}
\keyword{ANOVAShiny}
